+++ flutter/ios/Classes/SentryFlutterPluginApple.swift
@@ -179,6 +179,9 @@ public class SentryFlutterPluginApple: NSObject, FlutterPlugin {
         case "resumeAppHangTracking":
             resumeAppHangTracking(result)
 
+        case "nativeCrash":
+            crash()
+
         case "sendReplayForEvent":
 #if canImport(UIKit) && !SENTRY_NO_UIKIT && (os(iOS) || os(tvOS))
             PrivateSentrySDKOnly.captureReplay()
@@ -749,6 +752,10 @@ public class SentryFlutterPluginApple: NSObject, FlutterPlugin {
     private func resumeAppHangTracking(_ result: @escaping FlutterResult) {
         SentrySDK.resumeAppHangTracking()
         result("")
+    }
+
+    private func crash() {
+        SentrySDK.crash()
     }
 }